Refactors timeline component
