+++ moped-editor/src/views/dashboard/DashboardTimelineModal.js
@@ -1,13 +1,20 @@
 import React, { useState } from "react";
 import {
+  Box,
+  CardContent,
   Dialog,
   DialogTitle,
   DialogContent,
+  Grid,
   IconButton,
 } from "@material-ui/core";
 import CloseIcon from "@material-ui/icons/Close";
 import { makeStyles } from "@material-ui/core/styles";
-import ProjectTimeline from "../projects/projectView/ProjectTimeline";
+import { TIMELINE_QUERY } from "src/queries/project";
+import { useQuery } from "@apollo/client";
+import ApolloErrorHandler from "src/components/ApolloErrorHandler";
+import ProjectMilestones from "../projects/projectView/ProjectMilestones";
+import ProjectPhases from "../projects/projectView/ProjectPhases";
 
 const useStyles = makeStyles((theme) => ({
   dialogTitle: {
@@ -31,6 +38,18 @@ const DashboardTimelineModal = ({
 }) => {
   const classes = useStyles();
   const [isDialogOpen, setIsDialogOpen] = useState(false);
+
+  /**
+   * Queries Hook
+   * @type {boolean} - loading state
+   * @type {object} - details and messages when there is a query error
+   * @type {object} - data returned from Hasura
+   * @function refetch - Provides a manual callback to update the Apollo cache
+   * */
+  const { loading, error, data, refetch } = useQuery(TIMELINE_QUERY, {
+    variables: { projectId },
+    fetchPolicy: "no-cache",
+  });
 
   const handleDialogClose = () => {
     setIsDialogOpen(false);
@@ -58,10 +77,32 @@ const DashboardTimelineModal = ({
           </IconButton>
         </DialogTitle>
         <DialogContent>
-          <ProjectTimeline
-            table={table}
-            projectIdFromTimelineModal={projectId}
-          />
+          <ApolloErrorHandler error={error}>
+            <CardContent>
+              <Grid container spacing={2}>
+                <Grid item xs={12}>
+                  <Box mb={2} style={{ maxWidth: "100%" }}>
+                    {table === "phases" && (
+                      <ProjectPhases
+                        projectId={projectId}
+                        loading={loading}
+                        data={data}
+                        refetch={refetch}
+                      />
+                    )}
+                    {table === "milestones" && (
+                      <ProjectMilestones
+                        projectId={projectId}
+                        loading={loading}
+                        data={data}
+                        refetch={refetch}
+                      />
+                    )}
+                  </Box>
+                </Grid>
+              </Grid>
+            </CardContent>
+          </ApolloErrorHandler>
         </DialogContent>
       </Dialog>
     </>
